18MCA8095 Assignment Question (Android)
Create a number guessing game in android which will be able to suggest whether the number entered by the user is less than the number to be guessed or greater than the number to be guessed. Once the user enters the correct number, it should say “Yu hit the bull’s eye in N attempts”, where N is number of guesses take by user to guess the correct number.